# Search graph row index by commit hash using dictionary