@Radrik5 2016-09-03
Have you mesured memory overhead?