@vcpp 2016-09-10
No, I didn't measure memory overhead. I was irritated by performance, so I measured and optimized performance. Nice fact that we also avoided creating tons of linq iterators and anonymous type instances not more than a side effect.
Rather obvious that creating new objects in a loop with huge number of iterations causes frequent garbage collection, but it's not the issue itself. The issue is having such loop, especially in the UI thread.
Another thing which better not to do inside of such loop is thread synchronization, you can see 10^7 calls of the Monitor.Enter.