@EbenZhang 2017-01-03
Just FYI  in case any one finds similar issue on master branch.

When changing to another repo or after a fetch, the gitextensions of the leftpanel branch #3038 will crash because the row count (ui is loading) is 0 and the underlying graph data is ready thus returns a zero as the index and consequently causes a argument out of range exception. fixed in 374d15a2aea6e2131b99b656fd3f0b163edc21c2.